requirements.txt
pyusb==1.2.1
python-uinput==0.11.2